Added a lot of new stuff, added spring boot and working on moving to spring batch

+++ src/main/java/com/example/stackexchange/repo/BadgeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.stackexchange.entity.Badge;
 
-public interface BadgeRepository extends JpaRepository<Badge, Integer> {
+public interface BadgeRepository extends JpaRepository<Badge, Long> {
 
 }
